refactor(#3274): extract object-runtime enumeration/object-static helpers (WAVE-B slice 2)
Second slice of the `ensureObjectRuntime` decomposition, stacked on slice 1.
Lifts the enumeration + array-like-index + Object-static helper-build block
(~1,137 LOC) VERBATIM into a new sibling module
`object-runtime-enumeration.ts` (`buildObjectEnumerationHelpers`), threading the
captured scope through `ObjectEnumerationHelperState`.

Helpers relocated: __object_keys/__object_keys_forin,
__extern_length/__extern_get_idx/__extern_has_idx, __object_values/__object_entries,
__object_assign, __object_is.

object-runtime.ts: 7,720 -> 6,607 LOC; ensureObjectRuntime now ~3,813
(from ~7,378 pre-decomposition).

`externSetIdx` (the one function-scope const inside the region also read by the
later descriptor call) is re-derived immediately before that call — byte-neutral.
`buildExternGetIdxBody` is now exported and imported by the sibling (benign
function-decl cycle, call-time use only).

Gates: `prove-emit-identity check` -> IDENTICAL (39/39 gc/standalone/wasi);
`tsc --noEmit` -> 0; `tests/issue-3274-slice2.test.ts` (4) green — __object_is
(SameValue NaN/-0 edges) + __extern_length reachable + correct under
--target standalone, zero host imports.

Subtask of #3182. Stacked on #3076 (slice 1) — do not enqueue until slice 1 lands.

✓
